Refactor ConditionalListenerMixin and extract shared utilities
