+++ src/translations/en.json
@@ -7561,6 +7561,12 @@
                 "state_equal": "State is equal to",
                 "state_not_equal": "State is not equal to"
               },
+              "time": {
+                "label": "Time",
+                "after": "After",
+                "before": "Before",
+                "weekdays": "Weekdays"
+              },
               "location": {
                 "label": "Location",
                 "locations": "Locations",
